mv search text state to its own service out of the App container
