+++ catalog/app/containers/App/actions.js
@@ -1,11 +1,4 @@
-/* App actions */
+import { START } from './constants';
 
-import {
-  SET_SEARCH_TEXT,
-  START,
-} from './constants';
-
-
-export const setSearchText = (text) => ({ type: SET_SEARCH_TEXT, text });
 
 export const start = () => ({ type: START });
