+++ catalog/app/containers/App/constants.js
@@ -1,13 +1,5 @@
-/* App constants */
-import { fromJS } from 'immutable';
-
-
-export const initialState = fromJS({});
-
 export const DEFAULT_LOCALE = 'en';
 
 export const REDUX_KEY = 'app/App';
-
-export const SET_SEARCH_TEXT = 'app/App/SET_SEARCH_TEXT';
 
 export const START = 'app/App/START';
