--- catalog/app/containers/App/reducer.js
@@ -1,15 +1,0 @@
-/* App reducer */
-import {
-  SET_SEARCH_TEXT,
-  initialState,
-} from './constants';
-
-
-export default (state = initialState, action) => {
-  switch (action.type) {
-    case SET_SEARCH_TEXT:
-      return state.setIn(['searchText'], action.text);
-    default:
-      return state;
-  }
-};
