+++ catalog/app/containers/SearchResults/index.js
@@ -15,9 +15,8 @@ import Error from 'components/Error';
 import Help from 'components/Help';
 import Tag from 'components/Tag';
 import Working from 'components/Working';
-import { setSearchText } from 'containers/App/actions';
-import { selectSearchText } from 'containers/App/selectors';
 import { ES } from 'utils/AWS';
+import { setSearchText, selectSearchText } from 'utils/SearchProvider';
 import { composeComponent } from 'utils/reactTools';
 import { injectReducer } from 'utils/ReducerInjector';
 import { injectSaga } from 'utils/SagaInjector';
